hosttagprovider: Move struct
