+++ comp/metadata/host/hostimpl/hosttags/host_tag_provider_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package aggregator
+package hosttags
 
 import (
 	"testing"
